morning push

+++ BackEnd/src/main/java/com/example/ZIMNEW/Repository/MessageRepository.java
@@ -6,6 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.example.ZIMNEW.Model.Message;
 
 @Repository
-public interface MessageRepository extends JpaRepository<Message, Long> {
-
+public interface MessageRepository extends JpaRepository<Message, Integer> {
 }
